refactor(app-update): extract pending install task as direct RPC service
