+++ packages/kit-bg/src/services/ServiceAppUpdate.ts
@@ -2,8 +2,6 @@ import semver from 'semver';
 
 import { appApiClient } from '@onekeyhq/shared/src/appApiClient/appApiClient';
 import type {
-  IAppUpdateInstallTaskPayload,
-  IJsBundleSwitchTaskPayload,
   IPendingInstallTask,
   IResponseAppUpdateInfo,
 } from '@onekeyhq/shared/src/appUpdate';
@@ -18,7 +16,6 @@ import {
   backgroundMethod,
 } from '@onekeyhq/shared/src/background/backgroundDecorators';
 import { buildServiceEndpoint } from '@onekeyhq/shared/src/config/appConfig';
-import { OneKeyLocalError } from '@onekeyhq/shared/src/errors';
 import { ETranslations } from '@onekeyhq/shared/src/locale';
 import { defaultLogger } from '@onekeyhq/shared/src/logger/logger';
 import type { IUpdateDownloadedEvent } from '@onekeyhq/shared/src/modules3rdParty/auto-update';
@@ -34,37 +31,14 @@ import { EServiceEndpointEnum } from '@onekeyhq/shared/types/endpoint';
 
 import { appUpdatePersistAtom } from '../states/jotai/atoms';
 
-import {
-  clearPendingInstallTask,
-  getPendingInstallTask,
-  setPendingInstallTask,
-} from './pendingInstallTaskStorage';
 import ServiceBase from './ServiceBase';
+import ServicePendingInstallTask from './servicePendingInstallTask';
+import { PLACEHOLDER_SIGNATURE } from './servicePendingInstallTask';
 
 let syncTimerId: ReturnType<typeof setTimeout>;
 let downloadTimeoutId: ReturnType<typeof setTimeout>;
 let failedRecoveryTimerId: ReturnType<typeof setTimeout>;
 let firstLaunch = true;
-const PLACEHOLDER_SIGNATURE = 'dev-no-signature';
-const MAX_TASK_RETRY = 3;
-const MAX_FULL_FLOW_RETRY = 2;
-const MAX_RETRY_DELAY_MS = 10 * 60 * 1000;
-const RETRY_BASE_DELAY_MS = 30 * 1000;
-const RETRY_JITTER_MS = 5 * 1000;
-const TASK_FUSE_DURATION_MS = 24 * 60 * 60 * 1000;
-const RUNNING_TASK_STALE_MS = timerUtils.getTimeDurationMs({ minute: 5 });
-const PROCESS_LOCK_TIMEOUT_MS = timerUtils.getTimeDurationMs({ seconds: 30 });
-const IGNORED_TARGET_TTL_MS = timerUtils.getTimeDurationMs({ day: 30 });
-const FULL_FLOW_RETRY_TTL_MS = timerUtils.getTimeDurationMs({ day: 7 });
-const MAX_IGNORED_TARGETS = 50;
-const MAX_FULL_FLOW_RETRY_TARGETS = 100;
-const RETRY_TRIGGER_BUNDLE_MISSING = 'BUNDLE_MISSING';
-const RETRY_TRIGGER_VERIFY_FAILED = 'VERIFY_EXTRACTED_FAILED';
-const RETRY_TRIGGER_INTERRUPTED = 'INTERRUPTED';
-const TERMINAL_REASON_RETRY_EXHAUSTED = 'RETRY_EXHAUSTED';
-const TERMINAL_REASON_FULL_FLOW_RETRY_EXHAUSTED = 'FULL_FLOW_RETRY_EXHAUSTED';
-const PENDING_ACTION_SWITCH_BUNDLE = 'switch-bundle';
-const PENDING_ACTION_INSTALL_APP = 'install-app';
 
 function normalizeOptionalString(value: unknown): string | undefined {
   if (value === null || value === undefined) {
@@ -85,6 +59,12 @@ function normalizeOptionalNumber(value: unknown): number | undefined {
 class ServiceAppUpdate extends ServiceBase {
   constructor({ backgroundApi }: { backgroundApi: any }) {
     super({ backgroundApi });
+    if (!this.backgroundApi.servicePendingInstallTask) {
+      this.pendingInstallTaskServiceFallback = new ServicePendingInstallTask({
+        backgroundApi: this.backgroundApi,
+        refreshUpdateStatus: async () => this.refreshUpdateStatus(),
+      });
+    }
   }
 
   private isResetting = false;
@@ -93,11 +73,19 @@ class ServiceAppUpdate extends ServiceBase {
 
   cachedUpdateInfo: IResponseAppUpdateInfo | undefined;
 
-  private isProcessingPendingTask = false;
+  private pendingInstallTaskServiceFallback:
+    | ServicePendingInstallTask
+    | undefined;
 
-  private pendingTaskLockAcquiredAt = 0;
-
-  private fallbackRequestSeqCounter = 0;
+  private get pendingInstallTaskService() {
+    const service =
+      this.backgroundApi.servicePendingInstallTask ||
+      this.pendingInstallTaskServiceFallback;
+    if (!service) {
+      throw new Error('servicePendingInstallTask is not available');
+    }
+    return service;
+  }
 
   private logUpdateEvent(
     event: string,
@@ -126,262 +114,46 @@ class ServiceAppUpdate extends ServiceBase {
     };
   }
 
-  private async clearPendingTaskWithLog({
-    traceId,
-    requestSeq,
-    task,
-    clearReason,
-    level = 'info',
-  }: {
-    traceId: string;
-    requestSeq?: number;
-    task?: IPendingInstallTask | null;
-    clearReason: string;
-    level?: 'info' | 'warn' | 'error';
-  }) {
-    await clearPendingInstallTask();
-    this.logUpdateEvent(
-      'pending_task_cleared',
-      {
-        traceId,
-        requestSeq: requestSeq ?? null,
-        ...this.buildTaskLogFields(task),
-        clearReason,
-      },
-      level,
-    );
-  }
 
   private getTargetKey(taskOrTarget: {
     targetAppVersion: string;
     targetBundleVersion: string;
   }) {
-    return `${taskOrTarget.targetAppVersion}:${taskOrTarget.targetBundleVersion}`;
-  }
-
-  private isTargetAligned(
-    targetAppVersion: string,
-    targetBundleVersion: string,
-  ) {
-    return (
-      targetAppVersion === (platformEnv.version || '') &&
-      targetBundleVersion === String(platformEnv.bundleVersion || '')
-    );
-  }
-
-  private isTaskTargetAligned(task: IPendingInstallTask) {
-    if (task.type === 'appshell-install') {
-      return task.targetAppVersion === (platformEnv.version || '');
-    }
-    return this.isTargetAligned(task.targetAppVersion, task.targetBundleVersion);
+    return this.pendingInstallTaskService.getTargetKey(taskOrTarget);
   }
 
   private async nextRequestSeq() {
-    let requestSeq = 0;
-    await appUpdatePersistAtom.set((prev) => {
-      const prevSeq = Number(prev.lastRequestSeq || 0);
-      this.fallbackRequestSeqCounter += 1;
-      const nextSeq =
-        Number.isSafeInteger(prevSeq) && prevSeq > 0
-          ? prevSeq + 1
-          : Date.now() * 1000 + this.fallbackRequestSeqCounter;
-      requestSeq = nextSeq;
-      return {
-        ...prev,
-        lastRequestSeq: nextSeq,
-      };
-    });
-    this.fallbackRequestSeqCounter += 1;
-    return requestSeq || Date.now() * 1000 + this.fallbackRequestSeqCounter;
-  }
-
-  private pruneIgnoredTargets(
-    ignoredTargets: Record<
-      string,
-      { reason: string; createdAt: number; expiresAt: number }
-    >,
-    now: number,
-  ) {
-    const entries = Object.entries(ignoredTargets)
-      .filter(([, info]) => info.expiresAt > now)
-      .toSorted((a, b) => b[1].createdAt - a[1].createdAt)
-      .slice(0, MAX_IGNORED_TARGETS);
-    return Object.fromEntries(entries);
-  }
-
-  private pruneFullFlowRetryByTarget(
-    fullFlowRetryByTarget: Record<string, { count: number; updatedAt: number }>,
-    now: number,
-  ) {
-    const entries = Object.entries(fullFlowRetryByTarget)
-      .filter(([, info]) => now - info.updatedAt <= FULL_FLOW_RETRY_TTL_MS)
-      .toSorted((a, b) => b[1].updatedAt - a[1].updatedAt)
-      .slice(0, MAX_FULL_FLOW_RETRY_TARGETS);
-    return Object.fromEntries(entries);
+    return this.pendingInstallTaskService.nextRequestSeq();
   }
 
   private async cleanupUpdateControlState() {
-    const now = Date.now();
-    await appUpdatePersistAtom.set((prev) => ({
-      ...prev,
-      ignoredTargets: this.pruneIgnoredTargets(prev.ignoredTargets || {}, now),
-      fullFlowRetryByTarget: this.pruneFullFlowRetryByTarget(
-        prev.fullFlowRetryByTarget || {},
-        now,
-      ),
-    }));
+    await this.pendingInstallTaskService.cleanupUpdateControlState();
   }
 
-  private async resetTargetControlState(targetKey: string) {
-    await appUpdatePersistAtom.set((prev) => {
-      const nextIgnoredTargets = { ...prev.ignoredTargets };
-      delete nextIgnoredTargets[targetKey];
-      const nextFullFlowRetryByTarget = { ...prev.fullFlowRetryByTarget };
-      delete nextFullFlowRetryByTarget[targetKey];
-      return {
-        ...prev,
-        ignoredTargets: nextIgnoredTargets,
-        fullFlowRetryByTarget: nextFullFlowRetryByTarget,
-      };
-    });
-  }
-
-  private async incrementFullFlowRetry(targetKey: string) {
-    let nextCount = 0;
-    await appUpdatePersistAtom.set((prev) => {
-      const fullFlowRetryByTarget = {
-        ...prev.fullFlowRetryByTarget,
-      };
-      const current = fullFlowRetryByTarget[targetKey]?.count || 0;
-      nextCount = current + 1;
-      fullFlowRetryByTarget[targetKey] = {
-        count: nextCount,
-        updatedAt: Date.now(),
-      };
-      return {
-        ...prev,
-        fullFlowRetryByTarget,
-      };
-    });
-    return nextCount;
-  }
-
-  private async freezeAndIgnoreTarget(
+  private async shouldSkipTargetByControl(
     targetKey: string,
-    reason: string,
     traceId: string,
+    requestSeq: number,
+    emitLog = true,
   ) {
-    const now = Date.now();
-    await appUpdatePersistAtom.set((prev) => {
-      const ignoredTargets = this.pruneIgnoredTargets(
-        {
-          ...prev.ignoredTargets,
-          [targetKey]: {
-            reason,
-            createdAt: now,
-            expiresAt: now + IGNORED_TARGET_TTL_MS,
-          },
-        },
-        now,
-      );
-      return {
-        ...prev,
-        freezeUntil: now + TASK_FUSE_DURATION_MS,
-        ignoredTargets,
-      };
-    });
-    this.logUpdateEvent(
-      'update_control_frozen_or_ignored',
-      {
-        traceId,
-        target: targetKey,
-        freezeUntil: now + TASK_FUSE_DURATION_MS,
-        terminalReason: reason,
-      },
-      'error',
+    return this.pendingInstallTaskService.shouldSkipTargetByControl(
+      targetKey,
+      traceId,
+      requestSeq,
+      emitLog,
     );
   }
 
-  private isValidTaskBase(task: unknown): task is IPendingInstallTask {
-    if (!task || typeof task !== 'object') {
-      return false;
-    }
-    const t = task as IPendingInstallTask;
-    if (
-      !t.taskId ||
-      !Number.isFinite(t.revision) ||
-      ![PENDING_ACTION_SWITCH_BUNDLE, PENDING_ACTION_INSTALL_APP].includes(
-        t.action,
-      ) ||
-      !t.type ||
-      !t.targetAppVersion ||
-      !t.targetBundleVersion ||
-      !t.scheduledEnvAppVersion ||
-      !t.scheduledEnvBundleVersion ||
-      !Number.isFinite(t.createdAt) ||
-      !Number.isFinite(t.expiresAt) ||
-      !Number.isFinite(t.retryCount)
-    ) {
-      return false;
-    }
-    if (
-      !['pending', 'running', 'applied_waiting_verify', 'failed'].includes(
-        t.status,
-      )
-    ) {
-      return false;
-    }
-    if (!t.payload || typeof t.payload !== 'object') {
-      return false;
-    }
-    return true;
-  }
-
-  private isValidJsBundleSwitchPayload(payload: unknown) {
-    const p = payload as IJsBundleSwitchTaskPayload;
-    return !!(p?.appVersion && p.bundleVersion && p.signature);
-  }
-
-  private isValidAppShellInstallPayload(payload: unknown) {
-    const p = payload as IAppUpdateInstallTaskPayload;
-    if (!p?.latestVersion) {
-      return false;
-    }
-    if (p.channel !== 'direct' && p.channel !== 'store') {
-      return false;
-    }
-    if (p.channel === 'direct') {
-      return !!p.downloadUrl;
-    }
-    return !!p.storeUrl;
-  }
-
-  private isValidPendingInstallTask(
-    task: unknown,
-  ): task is IPendingInstallTask {
-    if (!this.isValidTaskBase(task)) {
-      return false;
-    }
-    if (task.type === 'jsbundle-switch') {
-      return (
-        task.action === PENDING_ACTION_SWITCH_BUNDLE &&
-        this.isValidJsBundleSwitchPayload(task.payload)
-      );
-    }
-    if (task.type === 'appshell-install') {
-      return (
-        task.action === PENDING_ACTION_INSTALL_APP &&
-        this.isValidAppShellInstallPayload(task.payload)
-      );
-    }
-    return false;
-  }
-
-  private getRetryDelayMs(retryCount: number) {
-    const expDelay = RETRY_BASE_DELAY_MS * 2 ** (retryCount - 1);
-    const baseDelay = Math.min(MAX_RETRY_DELAY_MS, expDelay);
-    const jitter = Math.floor(Math.random() * (RETRY_JITTER_MS + 1));
-    return baseDelay + jitter;
+  private async syncPendingInstallTaskWithReleaseInfo(args: {
+    releaseInfo: IResponseAppUpdateInfo | undefined;
+    requestSeq: number;
+    traceId: string;
+    stage: 'fetch' | 'ready_to_install';
+    appInfo?: Awaited<ReturnType<typeof appUpdatePersistAtom.get>>;
+  }) {
+    await this.pendingInstallTaskService.syncPendingInstallTaskWithReleaseInfo(
+      args,
+    );
   }
 
   private shouldUpdateFromReleaseInfo(releaseInfo: IResponseAppUpdateInfo) {
@@ -397,573 +169,6 @@ class ServiceAppUpdate extends ServiceBase {
       resolved.decision === 'jsBundleUpgrade' ||
       resolved.decision === 'jsBundleRollback'
     );
-  }
-
-  private buildPendingJsBundleTask(
-    releaseInfo: IResponseAppUpdateInfo,
-    revision: number,
-  ): IPendingInstallTask | undefined {
-    const appVersion = releaseInfo.version;
-    const bundleVersion = releaseInfo.jsBundleVersion;
-    if (!appVersion || !bundleVersion) {
-      return undefined;
-    }
-
-    const now = Date.now();
-    return {
-      taskId: `jsbundle:${appVersion}:${bundleVersion}`,
-      revision,
-      action: PENDING_ACTION_SWITCH_BUNDLE,
-      type: 'jsbundle-switch',
-      targetAppVersion: appVersion,
-      targetBundleVersion: bundleVersion,
-      scheduledEnvAppVersion: platformEnv.version || '',
-      scheduledEnvBundleVersion: String(platformEnv.bundleVersion || ''),
-      createdAt: now,
-      expiresAt: now + timerUtils.getTimeDurationMs({ day: 7 }),
-      retryCount: 0,
-      status: 'pending',
-      payload: {
-        appVersion,
-        bundleVersion,
-        signature: releaseInfo.jsBundle?.signature ?? PLACEHOLDER_SIGNATURE,
-      },
-    };
-  }
-
-  private buildPendingAppShellTask(
-    appInfo: Awaited<ReturnType<typeof appUpdatePersistAtom.get>>,
-    revision: number,
-  ): IPendingInstallTask | undefined {
-    if (!appInfo.latestVersion) {
-      return undefined;
-    }
-    const downloadUrl =
-      appInfo.downloadedEvent?.downloadUrl || appInfo.downloadUrl;
-    const downloadedFile = appInfo.downloadedEvent?.downloadedFile;
-    if (!downloadUrl || !downloadedFile) {
-      return undefined;
-    }
-    const now = Date.now();
-    return {
-      taskId: `appshell:${appInfo.latestVersion}:direct`,
-      revision,
-      action: PENDING_ACTION_INSTALL_APP,
-      type: 'appshell-install',
-      targetAppVersion: appInfo.latestVersion,
-      targetBundleVersion:
-        appInfo.jsBundleVersion || String(platformEnv.bundleVersion || ''),
-      scheduledEnvAppVersion: platformEnv.version || '',
-      scheduledEnvBundleVersion: String(platformEnv.bundleVersion || ''),
-      createdAt: now,
-      expiresAt: now + timerUtils.getTimeDurationMs({ day: 7 }),
-      retryCount: 0,
-      status: 'pending',
-      payload: {
-        latestVersion: appInfo.latestVersion,
-        updateStrategy: appInfo.updateStrategy,
-        channel: 'direct',
-        storeUrl: appInfo.storeUrl,
-        downloadUrl,
-        fileSize: appInfo.fileSize,
-        sha256: appInfo.downloadedEvent?.sha256,
-        signature: appInfo.downloadedEvent?.signature,
-      },
-    };
-  }
-
-  private async shouldSkipTargetByControl(
-    targetKey: string,
-    traceId: string,
-    requestSeq: number,
-    emitLog = true,
-  ) {
-    const now = Date.now();
-    const appInfo = await appUpdatePersistAtom.get();
-    if ((appInfo.freezeUntil || 0) > now) {
-      if (emitLog) {
-        this.logUpdateEvent('pending_task_upsert_decision', {
-          traceId,
-          requestSeq,
-          upsertAction: 'drop',
-          reason: 'frozen',
-          freezeUntil: appInfo.freezeUntil,
-          target: targetKey,
-        });
-      }
-      return true;
-    }
-    const ignored = appInfo.ignoredTargets?.[targetKey];
-    if (ignored && ignored.expiresAt > now) {
-      if (emitLog) {
-        this.logUpdateEvent('pending_task_upsert_decision', {
-          traceId,
-          requestSeq,
-          upsertAction: 'drop',
-          reason: 'ignored_target',
-          target: targetKey,
-        });
-      }
-      return true;
-    }
-    return false;
-  }
-
-  private async syncPendingInstallTaskWithReleaseInfo({
-    releaseInfo,
-    requestSeq,
-    traceId,
-    stage,
-    appInfo,
-  }: {
-    releaseInfo: IResponseAppUpdateInfo | undefined;
-    requestSeq: number;
-    traceId: string;
-    stage: 'fetch' | 'ready_to_install';
-    appInfo?: Awaited<ReturnType<typeof appUpdatePersistAtom.get>>;
-  }) {
-    if (!releaseInfo) {
-      return;
-    }
-
-    const decision = resolveUpdateDecision({
-      currentAppVersion: platformEnv.version,
-      currentBundleVersion: platformEnv.bundleVersion,
-      remoteAppVersion: releaseInfo.version,
-      remoteBundleVersion: releaseInfo.jsBundleVersion,
-      allowRollback: true,
-    });
-    this.logUpdateEvent('app_update_decision_resolved', {
-      traceId,
-      requestSeq,
-      decision: decision.decision,
-      reason: decision.reason,
-      allowRollback: true,
-      currentAppVersion: platformEnv.version,
-      currentBundleVersion: String(platformEnv.bundleVersion || ''),
-      targetAppVersion: releaseInfo.version ?? null,
-      targetBundleVersion: releaseInfo.jsBundleVersion ?? null,
-      stage,
-    });
-
-    if (
-      decision.decision !== 'appShellUpdate' &&
-      decision.decision !== 'jsBundleUpgrade' &&
-      decision.decision !== 'jsBundleRollback'
-    ) {
-      this.logUpdateEvent('pending_task_upsert_decision', {
-        traceId,
-        requestSeq,
-        upsertAction: 'noop',
-        reason: `decision_${decision.decision}`,
-        stage,
-      });
-      return;
-    }
-
-    if (releaseInfo.updateStrategy !== EUpdateStrategy.seamless) {
-      this.logUpdateEvent('pending_task_upsert_decision', {
-        traceId,
-        requestSeq,
-        upsertAction: 'drop',
-        reason: 'strategy_not_restart_install',
-        updateStrategy: releaseInfo.updateStrategy ?? null,
-        stage,
-      });
-      return;
-    }
-
-    if (stage !== 'ready_to_install') {
-      this.logUpdateEvent('pending_task_upsert_decision', {
-        traceId,
-        requestSeq,
-        upsertAction: 'drop',
-        reason: 'resources_not_ready',
-        stage,
-      });
-      return;
-    }
-
-    const runtimeAppInfo = appInfo || (await appUpdatePersistAtom.get());
-    if (!runtimeAppInfo.downloadedEvent?.downloadedFile) {
-      this.logUpdateEvent('pending_task_upsert_decision', {
-        traceId,
-        requestSeq,
-        upsertAction: 'drop',
-        reason: 'downloaded_event_missing',
-        stage,
-      });
-      return;
-    }
-
-    const task =
-      decision.decision === 'appShellUpdate'
-        ? this.buildPendingAppShellTask(runtimeAppInfo, requestSeq)
-        : this.buildPendingJsBundleTask(
-            {
-              ...releaseInfo,
-              jsBundle: {
-                ...releaseInfo.jsBundle,
-                signature:
-                  runtimeAppInfo.downloadedEvent?.signature ??
-                  releaseInfo.jsBundle?.signature,
-              },
-            },
-            requestSeq,
-          );
-    if (!task || !this.isValidPendingInstallTask(task)) {
-      this.logUpdateEvent(
-        'pending_task_upsert_decision',
-        {
-          traceId,
-          requestSeq,
-          upsertAction: 'drop',
-          reason: 'invalid_new_task',
-          stage,
-        },
-        'warn',
-      );
-      return;
-    }
-
-    const targetKey = this.getTargetKey(task);
-    if (await this.shouldSkipTargetByControl(targetKey, traceId, requestSeq)) {
-      return;
-    }
-
-    const existingRaw = await getPendingInstallTask();
-    if (!existingRaw) {
-      await setPendingInstallTask(task);
-      this.logUpdateEvent('pending_task_upsert_decision', {
-        traceId,
-        requestSeq,
-        taskId: task.taskId,
-        revision: task.revision,
-        action: task.action,
-        upsertAction: 'create',
-        reason: 'no_existing_task',
-        stage,
-      });
-      return;
-    }
-
-    if (!this.isValidPendingInstallTask(existingRaw)) {
-      this.logUpdateEvent(
-        'pending_task_unknown_type_dropped',
-        {
-          traceId,
-          requestSeq,
-          taskType: (existingRaw as { type?: string })?.type || 'unknown',
-        },
-        'warn',
-      );
-      await this.clearPendingTaskWithLog({
-        traceId,
-        requestSeq,
-        task: existingRaw as IPendingInstallTask,
-        clearReason: 'replace_invalid_task',
-        level: 'warn',
-      });
-      await setPendingInstallTask(task);
-      this.logUpdateEvent('pending_task_upsert_decision', {
-        traceId,
-        requestSeq,
-        taskId: task.taskId,
-        revision: task.revision,
-        action: task.action,
-        upsertAction: 'create',
-        reason: 'replace_invalid_task',
-        stage,
-      });
-      return;
-    }
-
-    const existing = existingRaw;
-    if (existing.revision > requestSeq) {
-      this.logUpdateEvent(
-        'pending_task_upsert_decision',
-        {
-          traceId,
-          requestSeq,
-          taskId: existing.taskId,
-          revision: existing.revision,
-          action: existing.action,
-          upsertAction: 'drop',
-          reason: 'stale_request_seq',
-          stage,
-        },
-        'warn',
-      );
-      return;
-    }
-
-    if (
-      existing.revision === requestSeq &&
-      existing.taskId === task.taskId &&
-      existing.action === task.action
-    ) {
-      this.logUpdateEvent('pending_task_upsert_decision', {
-        traceId,
-        requestSeq,
-        taskId: task.taskId,
-        revision: task.revision,
-        action: task.action,
-        upsertAction: 'noop',
-        reason: 'same_revision_same_target',
-        stage,
-      });
-      return;
-    }
-
-    if (existing.taskId === task.taskId) {
-      await setPendingInstallTask({
-        ...task,
-        retryCount: existing.retryCount,
-        nextRetryAt: existing.nextRetryAt,
-        status: existing.status,
-        runningStartedAt: existing.runningStartedAt,
-        lastError: existing.lastError,
-      });
-      this.logUpdateEvent('pending_task_upsert_decision', {
-        traceId,
-        requestSeq,
-        taskId: task.taskId,
-        revision: task.revision,
-        action: task.action,
-        upsertAction: 'update',
-        reason: 'same_target_keep_retry_state',
-        stage,
-      });
-      return;
-    }
-
-    await setPendingInstallTask(task);
-    this.logUpdateEvent('pending_task_upsert_decision', {
-      traceId,
-      requestSeq,
-      taskId: task.taskId,
-      revision: task.revision,
-      action: task.action,
-      upsertAction: 'update',
-      reason: 'newer_revision_replace_target',
-      stage,
-    });
-  }
-
-  private async markTaskFailed(
-    task: IPendingInstallTask,
-    message: string,
-    traceId: string,
-    requestSeq?: number,
-  ) {
-    const targetKey = this.getTargetKey(task);
-    const isFullFlowRetryTrigger =
-      message.includes(RETRY_TRIGGER_BUNDLE_MISSING) ||
-      message.includes(RETRY_TRIGGER_VERIFY_FAILED);
-
-    if (isFullFlowRetryTrigger) {
-      const fullFlowRetryCount = await this.incrementFullFlowRetry(targetKey);
-      this.logUpdateEvent(
-        'full_flow_retry_triggered',
-        {
-          traceId,
-          requestSeq: requestSeq ?? null,
-          taskId: task.taskId,
-          revision: task.revision,
-          action: task.action,
-          trigger: message.includes(RETRY_TRIGGER_BUNDLE_MISSING)
-            ? 'bundle_missing'
-            : 'verify_failed',
-          fullFlowRetryCount,
-          target: targetKey,
-        },
-        'warn',
-      );
-      if (fullFlowRetryCount > MAX_FULL_FLOW_RETRY) {
-        await this.clearPendingTaskWithLog({
-          traceId,
-          requestSeq,
-          task,
-          clearReason: 'full_flow_retry_exhausted',
-          level: 'warn',
-        });
-        await this.freezeAndIgnoreTarget(
-          targetKey,
-          TERMINAL_REASON_FULL_FLOW_RETRY_EXHAUSTED,
-          traceId,
-        );
-        return;
-      }
-      await this.clearPendingTaskWithLog({
-        traceId,
-        requestSeq,
-        task,
-        clearReason: 'full_flow_retry_fallback_to_refetch',
-        level: 'warn',
-      });
-      return;
-    }
-
-    const nextRetryCount = task.retryCount + 1;
-    const now = Date.now();
-    if (nextRetryCount >= MAX_TASK_RETRY) {
-      await this.clearPendingTaskWithLog({
-        traceId,
-        requestSeq,
-        task,
-        clearReason: 'switch_retry_exhausted',
-        level: 'warn',
-      });
-      await this.freezeAndIgnoreTarget(
-        targetKey,
-        TERMINAL_REASON_RETRY_EXHAUSTED,
-        traceId,
-      );
-      return;
-    }
-
-    const delayMs = this.getRetryDelayMs(nextRetryCount);
-    const nextRetryAt = now + delayMs;
-    await setPendingInstallTask({
-      ...task,
-      retryCount: nextRetryCount,
-      status: 'pending',
-      runningStartedAt: undefined,
-      lastError: message,
-      nextRetryAt,
-    });
-    this.logUpdateEvent(
-      'pending_retry_scheduled',
-      {
-        traceId,
-        requestSeq: requestSeq ?? null,
-        taskId: task.taskId,
-        revision: task.revision,
-        action: task.action,
-        retryCount: nextRetryCount,
-        nextRetryAt,
-        retryType:
-          message === RETRY_TRIGGER_INTERRUPTED
-            ? 'interrupted'
-            : task.action === PENDING_ACTION_INSTALL_APP
-              ? 'install'
-              : 'switch',
-      },
-      'warn',
-    );
-  }
-
-  private async executeBundleSwitchTask(
-    task: Extract<IPendingInstallTask, { type: 'jsbundle-switch' }>,
-  ) {
-    const payload = task.payload;
-    const { appVersion, bundleVersion, signature } = payload;
-    const bundleExists = await BundleUpdate.isBundleExists(
-      appVersion,
-      bundleVersion,
-    );
-    if (!bundleExists) {
-      throw new OneKeyLocalError(RETRY_TRIGGER_BUNDLE_MISSING);
-    }
-
-    try {
-      await BundleUpdate.verifyExtractedBundle(appVersion, bundleVersion);
-    } catch (error) {
-      await BundleUpdate.clearBundle();
-      throw new OneKeyLocalError(
-        `${RETRY_TRIGGER_VERIFY_FAILED}:${
-          (error as Error)?.message || 'unknown'
-        }`,
-      );
-    }
-
-    await BundleUpdate.switchBundle({
-      appVersion,
-      bundleVersion,
-      signature,
-    });
-  }
-
-  private async executeAppShellInstallTask(
-    task: Extract<IPendingInstallTask, { type: 'appshell-install' }>,
-  ) {
-    const payload = task.payload;
-    if (payload.channel !== 'direct') {
-      throw new OneKeyLocalError('APP_INSTALL_CHANNEL_UNSUPPORTED');
-    }
-    const appInfo = await appUpdatePersistAtom.get();
-    const downloadUrl = appInfo.downloadedEvent?.downloadUrl || payload.downloadUrl;
-    if (!appInfo.downloadedEvent?.downloadedFile || !downloadUrl) {
-      throw new OneKeyLocalError('APP_PACKAGE_MISSING');
-    }
-    await AppUpdate.installPackage({
-      ...appInfo,
-      latestVersion: payload.latestVersion,
-      updateStrategy: payload.updateStrategy,
-      storeUrl: payload.storeUrl || appInfo.storeUrl,
-      downloadUrl: payload.downloadUrl || appInfo.downloadUrl,
-      fileSize: payload.fileSize ?? appInfo.fileSize,
-      downloadedEvent: {
-        ...appInfo.downloadedEvent,
-        downloadUrl,
-      },
-    });
-  }
-
-  private async executePendingInstallTask(task: IPendingInstallTask) {
-    if (task.type === 'jsbundle-switch') {
-      await this.executeBundleSwitchTask(task);
-      return;
-    }
-    if (task.type === 'appshell-install') {
-      await this.executeAppShellInstallTask(task);
-      return;
-    }
-    const unknownType =
-      (task as unknown as { type?: string })?.type || 'unknown';
-    throw new OneKeyLocalError(`Unknown pending task type: ${unknownType}`);
-  }
-
-  private async runPostPendingRefresh({
-    traceId,
-    requestSeq,
-    task,
-  }: {
-    traceId: string;
-    requestSeq: number | null;
-    task?: Partial<IPendingInstallTask> | null;
-  }) {
-    const startedAt = Date.now();
-    this.logUpdateEvent('pending_post_process_refresh_start', {
-      traceId,
-      requestSeq,
-      ...this.buildTaskLogFields(task),
-    });
-    try {
-      await this.refreshUpdateStatus();
-      this.logUpdateEvent('pending_post_process_refresh_result', {
-        traceId,
-        requestSeq,
-        result: 'success',
-        durationMs: Date.now() - startedAt,
-        ...this.buildTaskLogFields(task),
-      });
-    } catch (error) {
-      const message = (error as Error)?.message ?? 'unknown';
-      this.logUpdateEvent(
-        'pending_post_process_refresh_result',
-        {
-          traceId,
-          requestSeq,
-          result: 'fail',
-          durationMs: Date.now() - startedAt,
-          errorCode: message,
-          errorMessage: message,
-          ...this.buildTaskLogFields(task),
-        },
-        'warn',
-      );
-    }
   }
 
   private startFailedRecoveryTimer() {
@@ -991,306 +196,7 @@ class ServiceAppUpdate extends ServiceBase {
 
   @backgroundMethod()
   async processPendingInstallTask() {
-    const traceId = generateUUID();
-    const requestSeq = null;
-    let shouldRunPostRefresh = false;
-    let processedTaskSnapshot: Partial<IPendingInstallTask> | null = null;
-    const lockNow = Date.now();
-    if (this.isProcessingPendingTask) {
-      const lockHeldMs = lockNow - this.pendingTaskLockAcquiredAt;
-      if (lockHeldMs > PROCESS_LOCK_TIMEOUT_MS) {
-        this.logUpdateEvent(
-          'pending_task_lock_state',
-          {
-            traceId,
-            requestSeq,
-            lockState: 'timeout',
-            lockHeldMs,
-            lockTimeoutMs: PROCESS_LOCK_TIMEOUT_MS,
-          },
-          'warn',
-        );
-        this.isProcessingPendingTask = false;
-        this.pendingTaskLockAcquiredAt = 0;
-      } else {
-        this.logUpdateEvent('pending_task_lock_state', {
-          traceId,
-          requestSeq,
-          lockState: 'reentrant',
-          lockHeldMs,
-          lockTimeoutMs: PROCESS_LOCK_TIMEOUT_MS,
-        });
-        return;
-      }
-    }
-
-    this.isProcessingPendingTask = true;
-    this.pendingTaskLockAcquiredAt = Date.now();
-    this.logUpdateEvent('pending_task_lock_state', {
-      traceId,
-      requestSeq,
-      lockState: 'acquired',
-      lockTimeoutMs: PROCESS_LOCK_TIMEOUT_MS,
-    });
-
-    try {
-      await this.cleanupUpdateControlState();
-
-      const rawTask = await getPendingInstallTask();
-      if (!rawTask) {
-        this.logUpdateEvent('pending_task_validation', {
-          traceId,
-          requestSeq,
-          isValid: false,
-          invalidReason: 'no_task',
-          ...this.buildTaskLogFields(null),
-        });
-        return;
-      }
-      shouldRunPostRefresh = true;
-      processedTaskSnapshot = rawTask as Partial<IPendingInstallTask>;
-
-      if (!this.isValidPendingInstallTask(rawTask)) {
-        this.logUpdateEvent(
-          'pending_task_unknown_type_dropped',
-          {
-            traceId,
-            requestSeq,
-            taskType: (rawTask as { type?: string })?.type || 'unknown',
-          },
-          'warn',
-        );
-        this.logUpdateEvent(
-          'pending_task_validation',
-          {
-            traceId,
-            requestSeq,
-            isValid: false,
-            invalidReason: 'invalid_task_payload',
-            ...this.buildTaskLogFields(rawTask as Partial<IPendingInstallTask>),
-          },
-          'warn',
-        );
-        await this.clearPendingTaskWithLog({
-          traceId,
-          task: rawTask as IPendingInstallTask,
-          clearReason: 'invalid_task_payload',
-          level: 'warn',
-        });
-        return;
-      }
-
-      let task = rawTask;
-      const now = Date.now();
-      this.logUpdateEvent('pending_task_validation', {
-        traceId,
-        requestSeq,
-        isValid: true,
-        invalidReason: null,
-        ...this.buildTaskLogFields(task),
-      });
-
-      if (task.expiresAt <= now) {
-        await this.clearPendingTaskWithLog({
-          traceId,
-          task,
-          clearReason: 'task_expired',
-          level: 'warn',
-        });
-        return;
-      }
-
-      const targetKey = this.getTargetKey(task);
-      if (task.status === 'failed') {
-        this.logUpdateEvent('pending_task_validation', {
-          traceId,
-          requestSeq,
-          isValid: true,
-          invalidReason: 'task_already_failed',
-          ...this.buildTaskLogFields(task),
-        });
-        return;
-      }
-
-      if (task.status === 'applied_waiting_verify') {
-        const aligned = this.isTaskTargetAligned(task);
-        this.logUpdateEvent(
-          'pending_verify_after_restart',
-          {
-            traceId,
-            requestSeq,
-            aligned,
-            currentAppVersion: platformEnv.version || '',
-            currentBundleVersion: String(platformEnv.bundleVersion || ''),
-            ...this.buildTaskLogFields(task),
-          },
-          aligned ? 'info' : 'error',
-        );
-        if (aligned) {
-          await this.resetTargetControlState(targetKey);
-          await this.clearPendingTaskWithLog({
-            traceId,
-            task,
-            clearReason: 'applied_task_verified_success',
-          });
-          return;
-        }
-        await this.markTaskFailed(
-          task,
-          'VERIFY_AFTER_RESTART_MISMATCH',
-          traceId,
-          undefined,
-        );
-        return;
-      }
-
-      if (task.status === 'running') {
-        const runningStartedAt = task.runningStartedAt || task.createdAt;
-        const runningDuration = now - runningStartedAt;
-        if (runningDuration <= RUNNING_TASK_STALE_MS) {
-          this.logUpdateEvent('pending_task_lock_state', {
-            traceId,
-            requestSeq,
-            lockState: 'reentrant',
-            runningDuration,
-            runningStaleMs: RUNNING_TASK_STALE_MS,
-            ...this.buildTaskLogFields(task),
-          });
-          return;
-        }
-        await this.markTaskFailed(
-          task,
-          RETRY_TRIGGER_INTERRUPTED,
-          traceId,
-          undefined,
-        );
-        const latestTask = await getPendingInstallTask();
-        if (!latestTask || !this.isValidPendingInstallTask(latestTask)) {
-          return;
-        }
-        task = latestTask;
-      }
-
-      const currentAppVersion = platformEnv.version || '';
-      const currentBundleVersion = String(platformEnv.bundleVersion || '');
-      const targetMatch = this.isTaskTargetAligned(task);
-      const scheduledMatch =
-        task.scheduledEnvAppVersion === currentAppVersion &&
-        task.scheduledEnvBundleVersion === currentBundleVersion;
-      let envMatch: 'target' | 'scheduled' | 'mismatch' = 'mismatch';
-      if (targetMatch) {
-        envMatch = 'target';
-      } else if (scheduledMatch) {
-        envMatch = 'scheduled';
-      }
-
-      this.logUpdateEvent('pending_task_env_check', {
-        traceId,
-        requestSeq,
-        envMatch,
-        currentAppVersion,
-        currentBundleVersion,
-        scheduledEnvAppVersion: task.scheduledEnvAppVersion,
-        scheduledEnvBundleVersion: task.scheduledEnvBundleVersion,
-        ...this.buildTaskLogFields(task),
-      });
-
-      if (targetMatch) {
-        await this.resetTargetControlState(targetKey);
-        await this.clearPendingTaskWithLog({
-          traceId,
-          task,
-          clearReason: 'target_already_aligned',
-        });
-        return;
-      }
-
-      if (!scheduledMatch) {
-        await this.clearPendingTaskWithLog({
-          traceId,
-          task,
-          clearReason: 'scheduled_env_mismatch',
-          level: 'warn',
-        });
-        return;
-      }
-
-      if (task.nextRetryAt && task.nextRetryAt > now) {
-        this.logUpdateEvent('pending_task_validation', {
-          traceId,
-          requestSeq,
-          isValid: true,
-          invalidReason: 'retry_backoff_not_elapsed',
-          ...this.buildTaskLogFields(task),
-        });
-        return;
-      }
-
-      const runningTask: IPendingInstallTask = {
-        ...task,
-        status: 'running',
-        runningStartedAt: Date.now(),
-      };
-      await setPendingInstallTask(runningTask);
-      const startedAt = Date.now();
-      this.logUpdateEvent('pending_switch_start', {
-        traceId,
-        requestSeq,
-        fromStatus: task.status,
-        ...this.buildTaskLogFields(runningTask),
-      });
-
-      try {
-        await this.executePendingInstallTask(runningTask);
-        const durationMs = Date.now() - startedAt;
-        await setPendingInstallTask({
-          ...runningTask,
-          status: 'applied_waiting_verify',
-          runningStartedAt: undefined,
-          lastError: undefined,
-        });
-        this.logUpdateEvent('pending_switch_result', {
-          traceId,
-          requestSeq,
-          result: 'success',
-          durationMs,
-          ...this.buildTaskLogFields(runningTask),
-        });
-      } catch (error) {
-        const durationMs = Date.now() - startedAt;
-        const message = (error as Error)?.message ?? 'unknown';
-        this.logUpdateEvent(
-          'pending_switch_result',
-          {
-            traceId,
-            requestSeq,
-            result: 'fail',
-            durationMs,
-            errorCode: message,
-            errorMessage: message,
-            ...this.buildTaskLogFields(runningTask),
-          },
-          'error',
-        );
-        await this.markTaskFailed(runningTask, message, traceId, undefined);
-      }
-    } finally {
-      if (shouldRunPostRefresh) {
-        await this.runPostPendingRefresh({
-          traceId,
-          requestSeq,
-          task: processedTaskSnapshot,
-        });
-      }
-      this.isProcessingPendingTask = false;
-      this.pendingTaskLockAcquiredAt = 0;
-      this.logUpdateEvent('pending_task_lock_state', {
-        traceId,
-        requestSeq,
-        lockState: 'released',
-        lockTimeoutMs: PROCESS_LOCK_TIMEOUT_MS,
-      });
-    }
+    await this.pendingInstallTaskService.processPendingInstallTask();
   }
 
   @backgroundMethod()
@@ -1928,13 +834,6 @@ class ServiceAppUpdate extends ServiceBase {
       },
       releaseInfo ? 'info' : 'warn',
     );
-    await this.syncPendingInstallTaskWithReleaseInfo({
-      releaseInfo,
-      requestSeq,
-      traceId,
-      stage: 'fetch',
-    });
-
     if (releaseInfo?.version || releaseInfo?.jsBundleVersion) {
       const decision = resolveUpdateDecision({
         currentAppVersion: platformEnv.version,
